+++ CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
--
+-   Extent button now zooms to site markers
+-   Map zooms to sites when filters are submitted
 
 ### Fixed